requirements.txt
groq
sqlalchemy
python-dotenv
